chore(common): Refactor and split code
