+++ src/get-cart.js
@@ -1,0 +1,7 @@
+export function getPrimaryCart() {
+    console.log('get primary cart');
+
+    return fetch("/api/storefront/carts/", {
+        "method": "GET"
+    })
+}
