+++ src/scripts.js
@@ -1,18 +1,7 @@
 console.log("Hello from your Chrome extension!");
 
-function getCookie(name) {
-    const value = `; ${document.cookie}`;
-    const parts = value.split(`; ${name}=`);
-    if (parts.length === 2) return parts.pop().split(';').shift();
-}
-
-async function getPrimaryCart() {
-    console.log('get primary cart');
-
-    return fetch("/api/storefront/carts/", {
-        "method": "GET"
-    })
-}
+import { getPrimaryCart } from './get-cart';
+import { getCookie } from './get-cookie';
 
 function getCheckout(checkoutId) {
     console.log('get primary checkout');
@@ -34,21 +23,6 @@ function addItemsToCart(cartId, items) {
         },
         "method": "POST",
         "body": `${items}`
-    })
-    .then(response => response.json())
-    .then(json => console.log(json))
-}
-
-function updateCheckoutMessage(cartId, message) {
-    console.log(`update checkout <${cartId}> message to <${message}>`);
-
-    fetch(`/api/storefront/checkouts/${cartId}`, {
-        "headers": {
-            'Content-Type': 'application/json',
-            'x-xsrf-token': getCookie('XSRF-TOKEN'),
-        },
-        "method": "PUT",
-        "body": `${message}`
     })
     .then(response => response.json())
     .then(json => console.log(json))
